Moving some more things around and renaming.
