+++ app/javascript/landing/components/applicationStatus.jsx
