+++ app/javascript/landing/index.jsx
@@ -3,7 +3,7 @@ import { render } from 'react-dom'
 import {Provider} from 'react-redux'
 import configureStore from './store/configureStore'
 import {fetchDate} from './actions/landing'
-import LandingContainer from "./components/landing/landingContainer";
+import LandingContainer from "./components/landingContainer";
 
 let store = configureStore({
   time: Date.now()
